% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{maxmin_cpp}
\alias{maxmin_cpp}
\title{Función para obtener el maxmimo de los minimos valores entre
las columna de una matriz y las filas de otra matriz, ambas cuadraticas}
\usage{
maxmin_cpp(matrix_1, matrix_2)
}
\arguments{
\item{NumericMatriz}{de tipo nxn.}
}
\description{
Función para obtener el maxmimo de los minimos valores entre
las columna de una matriz y las filas de otra matriz, ambas cuadraticas
}
\examples{
matriz_aleatoria <- function(n) {
matriz <- matrix(round(runif(n^2, 0, 1), 1), nrow = n, ncol = n)
 for (i in 1:n) {
   matriz[i, i] <- 1
 }
 return(matriz)
   }
set.seed(123)
A<- matrizAleatoria(3)
B<- matrizAleatoria(3)
#Donde A es
#      [,1] [,2] [,3]
#[1,]  1.0  0.9  0.5
#[2,]  0.8  1.0  0.9
#[3,]  0.4  0.0  1.0
#y B es
#     [,1] [,2] [,3]
#[1,]  1.0  0.7  0.9
#[2,]  1.0  1.0  0.2
#[3,]  0.5  0.1  1.0


maxmin_cpp(A,B)
#Da como resultado
#      [,1] [,2] [,3]
#[1,]  1.0  0.9  0.9
#[2,]  1.0  1.0  0.9
#[3,]  0.5  0.4  1.0

}
