% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mult_matrix_rcpp}
\alias{mult_matrix_rcpp}
\title{Función para realizar una multiplicacion matricial procesada en C++}
\usage{
mult_matrix_rcpp(matrix_1, matrix_2)
}
\arguments{
\item{NumericMatriz}{de tipo nxc.}
}
\description{
Función para realizar una multiplicacion matricial procesada en C++
}
\examples{
matriz_aleatoria <- function(n) {
matriz <- matrix(round(runif(n^2, 0, 1), 1), nrow = n, ncol = n)
 for (i in 1:n) {
   matriz[i, i] <- 1
 }
 return(matriz)
   }
set.seed(123)
A<- matrizAleatoria(3)
B<- matrizAleatoria(3)
#Donde A es
#      [,1] [,2] [,3]
#[1,]  1.0  0.9  0.5
#[2,]  0.8  1.0  0.9
#[3,]  0.4  0.0  1.0

#Donde B es
#      [,1] [,2] [,3]
#[1,]  1.0  0.9  0.5
#[2,]  1.0  1.0  0.9
#[3,]  0.5  0.1  1.


mult_matrix_rcpp(A,B)
#Da como resultado
#      [,1] [,2] [,3]
#[1,]  2.15 1.65 1.58
#[2,]  2.25 1.65 1.82
#[3,]  0.90 0.38 1.36

}
