% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rbind_rcpp}
\alias{rbind_rcpp}
\title{Función para unir una matriz con un vector y que este ultimo quede
en la ultima fila de la matriz.}
\usage{
rbind_rcpp(matrix_1, vector_1)
}
\arguments{
\item{NumericMatriz}{de tipo mxn.}

\item{NumericVector}{de largo n.}
}
\description{
Función para unir una matriz con un vector y que este ultimo quede
en la ultima fila de la matriz.
}
\examples{
matriz_aleatoria <- function(n) {
matriz <- matrix(round(runif(n^2, 0, 1), 1), nrow = n, ncol = n)
 for (i in 1:n) {
   matriz[i, i] <- 1
 }
 return(matriz)
   }
set.seed(123)
A<- matrizAleatoria(3)
B<- c(1, 2, 3)
#Donde A es
#      [,1] [,2] [,3]
#[1,]  1.0  0.9  0.5
#[2,]  0.8  1.0  0.9
#[3,]  0.4  0.0  1.0


rbind_rcpp(A,B)
#Da como resultado
#      [,1] [,2] [,3]
#[1,]  1.0  0.9  0.9
#[2,]  1.0  1.0  0.9
#[3,]  0.5  0.4  1.0
#[4,]    1   2     3

}
