% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{wich_min_rcpp}
\alias{wich_min_rcpp}
\title{Función para encontrar el o los minimos valores de una matriz, esta retorna en un data frame las posiciones (las posiciones estan pensadas para comenzar desde 1 como funciona en R).}
\usage{
wich_min_rcpp(matrix_1)
}
\arguments{
\item{NumericMatriz}{de tipo mxn.}
}
\description{
Función para encontrar el o los minimos valores de una matriz, esta retorna en un data frame las posiciones (las posiciones estan pensadas para comenzar desde 1 como funciona en R).
}
\examples{
matriz_aleatoria <- function(n) {
matriz <- matrix(round(runif(n^2, 0, 1), 1), nrow = n, ncol = n)
 for (i in 1:n) {
   matriz[i, i] <- 1
 }
 return(matriz)
   }
set.seed(123)
A<- matrizAleatoria(3)
#Donde A es
#      [,1] [,2] [,3]
#[1,]  1.0  0.9  0.5
#[2,]  0.8  1.0  0.9
#[3,]  0.4  0.0  1.0


wich_min_rcpp(A)
#Da como resultado
#  pos_max_fila pos_max_columna
#1            3               2

}
